admins: move purging command to new plugin
