+++ anjani/plugins/admins.py
@@ -25,7 +25,7 @@ from pyrogram.errors import (
     PeerIdInvalid,
     UserAdminInvalid,
     UserIdInvalid,
-    UserNotParticipant
+    UserNotParticipant,
 )
 from pyrogram.types import ChatMember, User
 
@@ -72,7 +72,7 @@ class Admins(plugin.Plugin):
 
         if not file:
             return await self.text(msg.chat.id, "gpic-no-photo")
-        
+
         await self.bot.client.set_chat_photo(msg.chat.id, photo=file.file_id)
 
     async def cmd_adminlist(self, ctx: command.Context) -> str:
@@ -84,8 +84,11 @@ class Admins(plugin.Plugin):
         async for member in self.bot.client.iter_chat_members(chat.id, filter="administrators"):  # type: ignore
             # Pyrogram is weird it returns all members even tho we provided the filter
             if member.status == "administrator":
-                name = (member.user.first_name + " " + member.user.last_name
-                        ) if member.user.last_name else member.user.first_name
+                name = (
+                    (member.user.first_name + " " + member.user.last_name)
+                    if member.user.last_name
+                    else member.user.first_name
+                )
                 admins += f"• [{name}](tg://user?id={member.user.id})\n"
 
         return admins
@@ -180,40 +183,6 @@ class Admins(plugin.Plugin):
 
         return await self.text(chat.id, "demote-success")
 
-    @command.filters(filters.can_delete)
-    async def cmd_del(self, ctx: command.Context) -> Optional[str]:
-        """Delete replied message"""
-        if not ctx.msg.reply_to_message:
-            return await self.text(ctx.chat.id, "error-reply-to-message")
-
-        await asyncio.gather(ctx.msg.reply_to_message.delete(),
-                             ctx.msg.delete())
-
-    @command.filters(filters.can_delete)
-    async def cmd_purge(self, ctx: command.Context) -> Optional[str]:
-        """purge message from message replied"""
-        if not ctx.msg.reply_to_message:
-            return await self.text(ctx.msg.chat.id, "error-reply-to-message")
-
-        time_start = datetime.now()
-        start, end = ctx.msg.reply_to_message.message_id, ctx.msg.message_id
-        messages = [*range(start, end)]
-
-        try:
-            await self.bot.client.delete_messages(chat_id=ctx.chat.id,
-                                                  message_ids=messages)
-        except MessageDeleteForbidden:
-            await ctx.respond(await self.text(ctx.chat.id, "purge-error", delete_after=5))
-            return
-        else:
-            await ctx.msg.delete()
-
-        time_end = datetime.now()
-        run_time = (time_end - time_start).seconds
-
-        await ctx.respond(await self.text(ctx.chat.id, "purge-done", len(messages), run_time), 
-                          delete_after=5)
-
     @command.filters(filters.can_restrict)
     async def cmd_kick(self, ctx: command.Context, user: Optional[User] = None) -> str:
         """Kick chat member"""
@@ -230,8 +199,9 @@ class Admins(plugin.Plugin):
             return await self.text(chat.id, "err-not-participant")
 
         await chat.kick_member(user.id)
-        ret, _ = await asyncio.gather(self.text(chat.id, "kick-done", user.first_name),
-                                      chat.unban_member(user.id))
+        ret, _ = await asyncio.gather(
+            self.text(chat.id, "kick-done", user.first_name), chat.unban_member(user.id)
+        )
 
         return ret
 
@@ -250,8 +220,9 @@ class Admins(plugin.Plugin):
         except UserNotParticipant:
             return await self.text(chat.id, "err-not-participant")
 
-        ret, _ = await asyncio.gather(self.text(chat.id, "ban-done", user.first_name),
-                                      chat.kick_member(user.id))
+        ret, _ = await asyncio.gather(
+            self.text(chat.id, "ban-done", user.first_name), chat.kick_member(user.id)
+        )
 
         return ret
